chore: moved convenience extension from Logging to GDSAnalytics so it is more widely available
