--- Sources/GDSAnalytics/Screens/NamedScreen.swift
@@ -1,6 +1,0 @@
-@available(*, deprecated, renamed: "ScreenType")
-public typealias NamedScreen = ScreenType
-
-public protocol ScreenType {
-    var name: String { get }
-}
