+++ Tests/GDSAnalyticsTests/Screens/ScreenTypeTests.swift
@@ -1,0 +1,13 @@
+@testable import GDSAnalytics
+import XCTest
+
+final class ScreenTypeTests: XCTestCase {
+    func testNameConformance() {
+        XCTAssertEqual(MockScreenType.drivingLicenceFrontInstructions.name,
+                       "drivingLicenceFrontInstructions")
+    }
+}
+
+enum MockScreenType: String, ScreenType {
+    case drivingLicenceFrontInstructions
+}
