--- Tests/LoggingTests/LoggingScreenTests.swift
@@ -1,13 +1,0 @@
-@testable import Logging
-import XCTest
-
-final class LoggingScreenTests: XCTestCase {
-    func testNameConformance() {
-        XCTAssertEqual(MockAnalyticsScreen.drivingLicenceFrontInstructions.name,
-                       "drivingLicenceFrontInstructions")
-    }
-}
-
-enum MockAnalyticsScreen: String, LoggableScreen {
-    case drivingLicenceFrontInstructions
-}
